+++ src/control/content/templates/locations.html
@@ -290,7 +290,13 @@
         </div>
         <hr class="none" />
         <div>
-          <h5 class="text-bold">Location Information & Images</h5>
+          <h5 class="text-bold">Location Images & Information</h5>
+          <hr class="none" />
+          <div class="clearfix">
+            <p class="info-note">The <b>list image</b> is displayed in list view only. To set a main image for the location
+              details page,
+              add images under <b>Location Images</b>. The first image will be used as the main image.</p>
+          </div>
         </div>
         <hr class="none" />
         <div class="form-group main clearfix">
