+++ src/control/design/index.html
@@ -193,3 +193,5 @@
   -->
 </body>
 </html>
+<script src="./design.controller.js"></script>
+<script src="./design.js"></script>
