+++ src/control/settings/settings.js
@@ -3,14 +3,14 @@
 // This is the entry point of your plugin's settings control.
 // Feel free to require any local or npm modules you've installed.
 import buildfire from 'buildfire';
-import Settings from '../../entities/Settings';
+import Settings from '../../widget/js/global/data/Settings';
 import authManager from '../../UserAccessControl/authManager';
 import SettingsController from "./settings.controller";
 import GlobalTagsListUI from './js/ui/globalTagsListUI';
 import GlobalEditorsListUI from './js/ui/globalEditorsListUI';
 import LocationEditingListUI from './js/ui/locationEditingListUI';
 import DialogComponent from './js/ui/dialog/dialog';
-import Locations from '../../repository/Locations';
+import Locations from '../../widget/js/global/repository/Locations';
 import { getDisplayName } from './js/util/helpers';
 import { isCameraControlVersion } from '../../shared/utils/mapUtils';
 import initLocationFields from './js/locationFields/controller';
@@ -503,7 +503,7 @@ const cropImage = (url, options) => {
   if (!url) {
     return "";
   }
-  return buildfire.imageLib.cropImage(url, options);
+  return buildfire.imageLib.cropImage(url, options) + '&crop=entropy';
 };
 
 const createEmptyHolder = (message = 'No Data', classes = '') => {
