+++ src/widget/index.html
@@ -44,3 +44,5 @@
 <section id="categories"></section>
 </body>
 </html>
+<script src="./widget.controller.js"></script>
+<script src="./widget.js"></script>
