+++ src/widget/js/global/constants/index.js
