+++ src/widget/js/global/constants/mapStyles.js
