+++ src/widget/js/global/data/Category.js
