+++ src/widget/js/global/data/Location.js
