+++ src/widget/js/global/helpers.js
