+++ src/widget/js/views/editView.js
@@ -1,12 +1,11 @@
 import state from '../state';
-import Location from '../../../entities/Location';
-import Locations from '../../../repository/Locations';
+import Location from '../global/data/Location';
+import Locations from '../global/repository/Locations';
 import DeepLink from '../../../utils/deeplink';
-import SearchEngine from '../../../repository/searchEngine';
+import SearchEngine from '../global/repository/searchEngine';
 import views from '../Views';
 import {
   createTemplate,
-  generateUUID,
   getDefaultOpeningHours,
   showToastMessage,
   transformCategoriesToText,
@@ -14,6 +13,7 @@ import {
   getActiveTemplate,
   cropImage,
 } from '../util/helpers';
+import { generateUUID } from '../global/helpers';
 import { uploadImages } from '../util/forms';
 import { navigateTo, resetBodyScroll } from '../util/ui';
 import Accordion from '../Accordion';
@@ -21,7 +21,7 @@ import { convertDateToTime, convertTimeToDate } from '../../../utils/datetime';
 import mapView from './mapView';
 import introView from './introView';
 import { validateOpeningHoursDuplication } from '../../../shared/utils';
-import constants from '../constants';
+import constants from '../global/constants';
 import accessManager from '../accessManager';
 import customFieldsController from './customFields';
